+++ src/lib.rs
@@ -54,7 +54,7 @@
 //! - Module `style`
 //!   - Colors - [`SetForegroundColor`](style/struct.SetForegroundColor.html),
 //!     [`SetBackgroundColor`](style/struct.SetBackgroundColor.html),
-//!     [`ResetColor`](style/struct.ResetColor.html)
+//!     [`ResetColor`](style/struct.ResetColor.html), [`SetColors`](style/struct.SetColors.html)
 //!   - Attributes - [`SetAttribute`](style/struct.SetAttribute.html), [`SetAttributes`](style/struct.SetAttributes.html),
 //!     [`PrintStyledContent`](style/struct.PrintStyledContent.html)
 //! - Module `terminal`
@@ -63,12 +63,14 @@
 //!   - Miscellaneous - [`Clear`](terminal/struct.Clear.html),
 //!     [`SetSize`](terminal/struct.SetSize.html)
 //!     [`SetTitle`](terminal/struct.SetTitle.html)
+//!     [`DisableLineWrap`](terminal/struct.DisableLineWrap.html)
+//!     [`EnableLineWrap`](terminal/struct.EnableLineWrap.html)
 //!   - Alternate screen - [`EnterAlternateScreen`](terminal/struct.EnterAlternateScreen.html),
 //!     [`LeaveAlternateScreen`](terminal/struct.LeaveAlternateScreen.html)
 //!
 //! ### Command Execution
 //!
-//! There are two different way's to execute commands:
+//! There are two different ways to execute commands:
 //!
 //! * [Lazy Execution](#lazy-execution)
 //! * [Direct Execution](#direct-execution)
@@ -248,7 +250,8 @@ pub mod tty;
 
 mod ansi;
 #[cfg(windows)]
-pub(crate) mod ansi_support;
+/// A module that exposes one function to check if the current terminal supports ansi sequences.
+pub mod ansi_support;
 mod command;
 mod error;
 pub(crate) mod macros;