+++ src/style/enums.rs
@@ -1,6 +1,0 @@
-pub(crate) use self::colored::Colored;
-pub use self::{attribute::Attribute, color::Color};
-
-mod attribute;
-mod color;
-mod colored;